services/webapp_bak/requirements.txt
tweepy==3.8.0
Click==7.0
dash==1.7.0
dash-core-components==1.6.0
dash-html-components==1.0.*
dash-renderer==1.2.2
dash-table==4.5.1
Jinja2==2.10.1
MarkupSafe==1.1.*
plotly==4.1.1
PyYAML==5.1.2
retrying==1.3.3
six==1.12.0
Werkzeug==0.15.*
pandas==0.25.*
nltk==3.4.*
textblob==0.15.*
psycopg2-binary==2.8.*
SQLAlchemy==1.3.12
flask_sqlalchemy==2.*
Flask-Migrate==2.5.*
Flask-Login==0.4.*
pandas-datareader==0.8.0
Flask-WTF==0.14.*
wrapt==1.11.2
WTForms==2.2.1
Flask-SocketIO==4.2.*
celery==4.4.*
Pattern==3.6.*
redis==3.3.11
